# expression: implement vectorized evaluation for `builtinTimestamp2ArgsSig`