<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
implement vectorized evaluation for `builtinTimestamp2ArgsSig`
To: #12101 

### What is changed and how it works?
```
BenchmarkVectorizedBuiltinTimeFunc/builtinTimestamp2ArgsSig-VecBuiltinFunc-4                 411       2993111 ns/op   298692 B/op      15195 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinTimestamp2ArgsSig-NonVecBuiltinFunc-4              385       2952825 ns/op   298370 B/op      15195 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinTimestamp2ArgsSig-VecBuiltinFunc#01-4             1519        878661 ns/op   117043 B/op       5635 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinTimestamp2ArgsSig-NonVecBuiltinFunc#01-4          1407        964736 ns/op   148936 B/op       7168 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinTimestamp2ArgsSig-VecBuiltinFunc#02-4              282       3801798 ns/op   484416 B/op      10933 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinTimestamp2ArgsSig-NonVecBuiltinFunc#02-4           266       4304289 ns/op   484416 B/op      10933 allocs/op

```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test